chore: bring branch into pre-commit compliance
Changes needed for the eager-init PR's pre-commit run to land green:

- Apply ruff format to existing files that hadn't been kept in sync
  (api_server, steering registry, request, two model files, and three
  steering tests).
- Cast any_layer.steering_index to torch.Tensor in the
  steering mixin so mypy can index it; the old code worked at runtime
  but the type was `Tensor | Module`. Drop the inner-typed
  dict[int, nn.Module] annotation in _init_steering_state to
  match the long-standing untyped pattern used elsewhere in the file
  (mypy can't narrow mod.layer_idx from a hasattr check).
- Add the vllm/third_party/** glob to the ruff format exclude so
  we don't reformat third-party vendored code (the ruff lint section
  already excluded it).
- Exclude vllm/model_executor/layers/fused_moe/runner/moe_runner_base.py
  from the mypy pre-commit script — its signature mismatches with
  MoERunner predate this work.
- Add upstream/RFC.md and upstream/prior.md to the typos
  exclude (false-positives from the `**X**word` markdown bold
  abbreviation pattern), and exclude `upstream/*.md` from
  markdownlint (prior-art reference docs we don't maintain to vLLM
  doc style).
- Drop a duplicated 'Generative Scoring API' section in
  docs/serving/openai_compatible_server.md (merge artifact, second
  copy had a broken in-page link to a non-existent #score-api anchor).
- Add a language tag to the one fenced code block in README.md.
- Regenerate docs/design/attention_backends.md via the existing
  generator hook (10.x → 10.0).

+++ tests/model_executor/layers/test_steering_op.py
@@ -30,9 +30,7 @@ def _apply_steering(
 ) -> torch.Tensor:
     """Reference implementation of the indexed-gather steering math."""
     N = hidden_states.shape[0]
-    return hidden_states + steering_table[steering_index[:N]].to(
-        hidden_states.dtype
-    )
+    return hidden_states + steering_table[steering_index[:N]].to(hidden_states.dtype)
 
 
 class TestIndexedGatherSteering:
